Fix CI: Remove --install-option.
This is deprecated in pip 23 which we switched to recently.  --install option was passed to setuptools, but pip is now using wheel install only from what I understand. --pre signals using pre-releases when available, and it's also a pip option. Lets pass it directly to pip and see what happens.

More info:
https://github.com/pypa/pip/issues/11358
https://discuss.python.org/t/passing-command-line-arguments-to-pip-install-after-install-options-deprecation/22981/6