Add new registry setting to disable or enable the Plone toolbar.
The new setting `plone.app.mosaic.settings.disable_edit_bar` allows to show or
hide the Plone toolbar. The default is `True` which hides the toolbar when the
Mosaic editor is loaded. There are usecases where you want to provide a
customized edit bar to all users and not hide it. This setting allows to show
the toolbar.